host/containerinit: Wait 2 seconds before forwarding SIGTERM/INT
We also now close the heartbeaters synchronously.

The purpose of this change is to give the router additional grace
period to remove backends that will be shutting down shortly from
it's routing table.
This ensures that new client requests won't be sent on dead or
soon to be dead connections during scale down events, like during
a deploy for instance.

Signed-off-by: Joseph Glanville <jpg@jpg.id.au>