discoverd: Fix closing the heartbeater on shutdown
The heartbeater uses the HTTP API, so shutting down the API before
closing the heartbeater leads to a "discoverd: shutting down" error, and
since the error is a retriable one, the heartbeat close blocks for 10s.

When this happens on the leader during a deployment, it causes all
service heartbeats to expire, wreaking havoc on the cluster.

Signed-off-by: Lewis Marshall <lewis@lmars.net>