controller: Replace ProcessType.Data with ProcessType.Volumes
This gives us more flexibility for specifying extra configuration for
volume provisioning (e.g. indicating that the volume should be deleted
when the job stops).

I have introduced a VolumeReq type (rather than just Volume) because in
the future we want to model actual volumes in the controller, whereas
ProcessType.Volumes are just requests for volumes to be provisioned.

Signed-off-by: Lewis Marshall <lewis@lmars.net>