bootstrap: Move SINGLETON flag on db appliances to app
Fixes Sirenia scale up code for singleton clusters.
Without this change the formation will be scaled to 3 but the state
machine will start in singleton mode and freeze the cluster with only 1
active member but 2 extra unassigned instances.
This change ensures on scale up only 1 database instance will be
started.

Signed-off-by: Joseph Glanville <jpg@jpg.id.au>