all: Reuse HTTP client transports
This fixes a connection leak in the cluster-monitor fixer, which
creates new clients for some phases. These new clients were not
subsequently re-used and the established connections in the HTTP
connection pool were leaking.

The solution is to reuse the same HTTP transport across all
instances of all clients, which this change implements.

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>