router: Add support for inbound PROXY protocol v1
This adds support for version 1 of the PROXY protocol to indicate
the source IP for incoming HTTP and HTTPS connections.

http://www.haproxy.org/download/1.5/doc/proxy-protocol.txt

The feature is off by default, and can be enabled by setting the
environment variable `PROXY_PROTOCOL=true` while running
a non-backup `flynn-host bootstrap` or configuring the router
directly after bootstrap with the command:

    flynn -a router env set PROXY_PROTOCOL=true

This feature is useful to integrate external load balancers in TCP
mode like the AWS Classic Load Balancer (previously called ELB) and
HAProxy.

https://docs.aws.amazon.com/elasticloadbalancing/latest/classic/enable-proxy-protocol.html
https://cbonte.github.io/haproxy-dconv/1.6/configuration.html#5.2-send-proxy

References #68

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>