bootstrap: Just wait for mariadb / mongodb API
The /ping endpoint tries to determine the status of the sirenia cluster,
which is not running for either mariadb or mongodb at bootstrap time, so
the request times out.

This used to fail immediately, but now fails after 30s since switching
HTTP transports in #3698 (which is a correct fix, if you want to know
the status of the sirenia cluster, you should be retrying dial errors).

Signed-off-by: Lewis Marshall <lewis@lmars.net>