host: Prevent deadlock when restoring state
Container cleanup emits a cleanup event which attempts to persist to the
state database, but the state database is open for reading when
restoring state during startup, so running cleanup can deadlock.

Container.Watch (which is run in a goroutine during restore) runs the
cleanup in a deferred call, so there is no need to run the cleanup
directly in UnmarshalState, thus preventing the deadlock.

Signed-off-by: Lewis Marshall <lewis@lmars.net>