scheduler: Only sync active jobs from hosts
Loading all jobs from each host in the cluster leads to ever increasing
network i/o and load in the cluster as the number of stopped jobs
increases.

Once the scheduler knows a job has stopped, there is no need to sync it
again, so the scheduler has been updated to just sync the starting /
running jobs from each host (which should be a small number of jobs) and
determine which jobs have stopped by their absence from that list.

Signed-off-by: Lewis Marshall <lewis@lmars.net>