test: Use a longer shutdown delay in host update tests
When updating a host running in a nested cluster, the whole host dies
when the original flynn-host daemon exits (because it is PID 1 in the
container) so all jobs on that host also die, leading to test failures
like "unexpected EOF".

The delay before exiting can now be specified by the client to give the
tests more time to make assertions.

Signed-off-by: Lewis Marshall <lewis@lmars.net>