discoverd: Use a low HTTP client timeout
Not having a timeout meant that, after a successful dial to a discoverd
server, subsequent HTTP requests could hang indefinitely, which is
particular problematic for heartbeats as then the service will get
marked as down even though the job is still running, so apps will
become unavailable.

Reduces the need for #3792 and #3766 as services should not flap like
they used to when the network becomes unreliable.

Fixes #3953
Fixes #3748
Fixes #3680
Fixes #3313

Signed-off-by: Lewis Marshall <lewis@lmars.net>