cli: Fix pre-ScaleRequest formation scaling
In an effort to fix #3983, a bug that affected versions older than
v20170309.0, 6a298e8d64fb2818e6704f85d542576c2e593175 was committed
which introduced a new bug for the same Flynn versions where process
types that are not specified in the `flynn scale` command are scaled
to zero because the existing formation map is not copied over.

This patch fixes the issue by ensuring that unspecified process
types are copied over properly and not modified by `flynn scale`.

Closes #4016

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>