LICENSE: edit Flynn license info so that GitHub recognizes it
GitHub uses a library called Licensee to identify a project's license
type. It shows this information in the status bar and via the API if it
can unambiguously identify the license.

This commit modifies LICENSE so that it matches the standard format of
the BSD 3-Clause license. It also updates the README with a new
"Licensing" section that provides additional clarity regarding the
project's license type and copyright holder, along with a link to the
LICENSE file.

Collectively, these changes allow Licensee to successfully identify the
license type of Flynn as BSD 3-Clause.

Signed-off-by: Andrea Kao <eirinikos@gmail.com>