router: Ensure that drain_backends is consistent for service
If there are multiple routes that point at the same service, and some
have drain_backends disabled, while others have drain_backends enabled,
all routes will have the flag set to a single non-deterministic value in
the router service cache, which can result in routes not being drained
properly.

This patch solves the problem by toggling any existing routes with
mismatched drain_backends to true, and adding a database check so that
new routes cannot be created with this inconsistency.

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>