builder: Downgrade host image to ubuntu-xenial
The bionic ZFS packages are not compatible with the older xenial
packages used in CI (due to a bug in recent versions of ZFS using
backing files), and so the host image needs to be based off xenial so
that its ZFS package can be used to create zpools when starting test
clusters.

This has no effect on the released flynn-host as the binary will
continue to be built with the Go image (that is based on bionic)
and the host image is only used for testing.

Signed-off-by: Lewis Marshall <lewis@lmars.net>